+++ KenChainer/Core/NumberChainProblem.cs
@@ -2,12 +2,12 @@
 {
     public class NumberChainProblem
     {
-        public NumberChainProblem(ushort result, params ushort[] numbers)
+        public NumberChainProblem(ushort result, params byte[] numbers)
         {
             Numbers = numbers;
             Result = result;
         }
-        public ushort[] Numbers { get; private set; }
+        public byte[] Numbers { get; private set; }
         public ushort Result { get; private set; }
     }
 }
